Renamed TokenDisplay to TokenListItem to follow the convention
