+++ src/popup/router/components/FungibleTokens/TokensListItem.vue
